requirements.txt
django==4.2.5
psycopg==3.1.14
django-filter==23.4
django-environ==0.11.2
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0